+++ src/components/input/Select/index.tsx
@@ -1,13 +1,4 @@
-import type {
-  BaseSyntheticEvent,
-  ChangeEvent,
-  ChangeEventHandler,
-  FocusEvent,
-  InputHTMLAttributes,
-  KeyboardEventHandler,
-  MouseEvent,
-  ReactNode,
-  RefObject,
+import {
   Children,
   cloneElement,
   forwardRef,
@@ -22,7 +13,7 @@ import styled, { useTheme } from 'styled-components';
 
 import { LIGHT_THEME } from '#src/components/themes';
 import { OpenStatusButton } from '#src/components/OpenStatusButton';
-import { keyboardKey } from '../../common/keyboardKey';
+import { keyboardKey } from '#src/components/common/keyboardKey.js';
 import { refSetter } from '#src/components/common/utils/refSetter';
 import { InputIconButton } from '#src/components/InputIconButton';
 import { ReactComponent as CloseOutlineSvg } from '@admiral-ds/icons/build/service/CloseOutline.svg';
@@ -75,10 +66,10 @@ export const DropDownText = styled(OptionWrapper)`
 type PartialOption = { value: string; disabled: boolean } & IConstantOption;
 const findAbledOptionValue = (options: PartialOption[]) => options.find(({ disabled }) => !disabled)?.value;
 
-const stopPropagation = (evt: BaseSyntheticEvent) => evt.stopPropagation();
+const stopPropagation = (evt: React.BaseSyntheticEvent) => evt.stopPropagation();
 
 export interface SelectProps
-  extends Omit<InputHTMLAttributes<HTMLSelectElement>, 'onFocus' | 'onBlur'>,
+  extends Omit<React.InputHTMLAttributes<HTMLSelectElement>, 'onFocus' | 'onBlur'>,
     DropContainerStyles,
     Pick<DropMenuComponentProps, 'targetElement' | 'renderTopPanel' | 'renderBottomPanel'> {
   value?: string | string[];
@@ -120,21 +111,21 @@ export interface SelectProps
    * Взамен используйте параметр targetElement.
    *
    * Референс на контейнер для правильного позиционирования выпадающего списка */
-  portalTargetRef?: RefObject<HTMLElement>;
+  portalTargetRef?: React.RefObject<HTMLElement>;
 
   /** Ref внутреннего input компонента */
-  inputTargetRef?: RefObject<HTMLInputElement>;
+  inputTargetRef?: React.RefObject<HTMLInputElement>;
 
   /** Делает высоту компонента больше или меньше обычной */
   dimension?: ComponentDimension;
 
   /** Иконки для отображения в правом углу поля */
-  icons?: ReactNode;
+  icons?: React.ReactNode;
 
   /** Статус поля */
   status?: InputStatus;
 
-  renderSelectValue?: (value: string | string[] | undefined, searchText: string) => ReactNode;
+  renderSelectValue?: (value: string | string[] | undefined, searchText: string) => React.ReactNode;
 
   /**  Значение введенное пользователем для поиска */
   inputValue?: string;
@@ -142,11 +133,11 @@ export interface SelectProps
   /** первоначальное значение в строке поиска без переведения строки в контролируемый компонент */
   defaultInputValue?: string;
 
-  onInputChange?: ChangeEventHandler<HTMLInputElement>;
+  onInputChange?: React.ChangeEventHandler<HTMLInputElement>;
 
-  onFocus?: (evt: FocusEvent<HTMLDivElement>) => void;
+  onFocus?: (evt: React.FocusEvent<HTMLDivElement>) => void;
 
-  onBlur?: (evt: FocusEvent<HTMLDivElement>) => void;
+  onBlur?: (evt: React.FocusEvent<HTMLDivElement>) => void;
 
   /**
    * @deprecated Помечено как deprecated в версии 8.10.0, будет удалено в 10.x.x версии.
@@ -163,7 +154,7 @@ export interface SelectProps
    **/
   locale?: {
     /** Сообщение, отображаемое при пустом наборе опций */
-    emptyMessage?: ReactNode;
+    emptyMessage?: React.ReactNode;
   };
 
   /**
@@ -172,7 +163,7 @@ export interface SelectProps
    *
    * Позволяет добавить панель внизу под выпадающим списком
    **/
-  renderDropDownBottomPanel?: (props: RenderPanelProps) => ReactNode;
+  renderDropDownBottomPanel?: (props: RenderPanelProps) => React.ReactNode;
 
   /**
    * @deprecated Помечено как deprecated в версии 8.10.0, будет удалено в 10.x.x версии.
@@ -180,7 +171,7 @@ export interface SelectProps
    *
    * Позволяет добавить панель сверху над выпадающим списком
    **/
-  renderDropDownTopPanel?: (props: RenderPanelProps) => ReactNode;
+  renderDropDownTopPanel?: (props: RenderPanelProps) => React.ReactNode;
 
   /** Состояние принудительного открытия выпадающего списка опций */
   forcedOpen?: boolean;
@@ -189,16 +180,16 @@ export interface SelectProps
   onChangeDropDownState?: (opened: boolean) => void;
 
   /** Inner input keyboard event handler */
-  onInputKeyDown?: KeyboardEventHandler<HTMLInputElement>;
+  onInputKeyDown?: React.KeyboardEventHandler<HTMLInputElement>;
 
   /** Inner input keyboard event handler */
-  onInputKeyUp?: KeyboardEventHandler<HTMLInputElement>;
+  onInputKeyUp?: React.KeyboardEventHandler<HTMLInputElement>;
 
   /** Inner input keyboard event handler */
-  onInputKeyUpCapture?: KeyboardEventHandler<HTMLInputElement>;
+  onInputKeyUpCapture?: React.KeyboardEventHandler<HTMLInputElement>;
 
   /** Inner input keyboard event handler */
-  onInputKeyDownCapture?: KeyboardEventHandler<HTMLInputElement>;
+  onInputKeyDownCapture?: React.KeyboardEventHandler<HTMLInputElement>;
 
   /** Данная опция позволяет при фильтрации искать по строке целиком или по отдельным словам */
   searchFormat?: SearchFormat;
@@ -542,13 +533,13 @@ export const Select = forwardRef<HTMLSelectElement, SelectProps>(
       setIsSearchPanelOpen((prev) => !prev);
     };
 
-    const mutateAndExtendTargetInputValue = (evt: ChangeEvent<HTMLInputElement>) => {
+    const mutateAndExtendTargetInputValue = (evt: React.ChangeEvent<HTMLInputElement>) => {
       if (!mutableState.current.shouldExtendInputValue || !visibleValueIsString) return;
       evt.target.value = `${visibleValue}${evt.target.value}`;
       mutableState.current.shouldExtendInputValue = false;
     };
 
-    const onLocalInputChange = (evt: ChangeEvent<HTMLInputElement>) => {
+    const onLocalInputChange = (evt: React.ChangeEvent<HTMLInputElement>) => {
       if (!multiple) setShouldRenderSelectValue(false);
       mutateAndExtendTargetInputValue(evt);
       if (inputValue === undefined) setSearchValue(evt.target.value);
@@ -648,14 +639,14 @@ export const Select = forwardRef<HTMLSelectElement, SelectProps>(
       };
     }, []);
 
-    const onFocus = (evt: FocusEvent<HTMLDivElement>) => {
+    const onFocus = (evt: React.FocusEvent<HTMLDivElement>) => {
       if (!isFocused) {
         setIsFocused(true);
         onFocusFromProps?.(evt);
       }
     };
 
-    const handleWrapperBlur = (evt: FocusEvent<HTMLDivElement>) => {
+    const handleWrapperBlur = (evt: React.FocusEvent<HTMLDivElement>) => {
       // если фокус переходит не на инпут, содержащийся внутри компонента
       if (!evt.currentTarget.contains(evt.relatedTarget) && !dropDownRef.current?.contains(evt.relatedTarget)) {
         setIsFocused(false);
@@ -665,7 +656,7 @@ export const Select = forwardRef<HTMLSelectElement, SelectProps>(
       }
     };
 
-    const handleNativeControlChange = (evt: ChangeEvent<HTMLSelectElement>) => {
+    const handleNativeControlChange = (evt: React.ChangeEvent<HTMLSelectElement>) => {
       if (isKeyboardEvent.current && modeIsSelect) {
         setPreseleceted(evt.target.value);
         return;
@@ -734,7 +725,7 @@ export const Select = forwardRef<HTMLSelectElement, SelectProps>(
       if (!selectIsUncontrolled) setSelectedValue(value);
     }, [value, selectIsUncontrolled]);
 
-    const handleWrapperClick = (e: MouseEvent) => {
+    const handleWrapperClick = (e: React.MouseEvent) => {
       if (e.target && dropDownRef.current?.contains(e.target as Node)) return;
 
       const passClick = !modeIsSelect && isSearchPanelOpen;